Refactor command class hierarchy to remove an unnecessary intermediate base class Command.
